% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2G_static_function.R
\name{G2G_static_LL}
\alias{G2G_static_LL}
\title{Fit G2G Model with Static Covariates}
\usage{
G2G_static_LL(par, df)
}
\arguments{
\item{par}{Parameter vector}

\item{df}{Data frame with time, status, and covariates}
}
\value{
Negative log-likelihood value
}
\description{
Fit G2G Model with Static Covariates
}
\keyword{internal}
