% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2Gcov-package.R
\docType{package}
\name{G2Gcov-package}
\alias{G2Gcov}
\alias{G2Gcov-package}
\title{G2Gcov: Grassia(II)-Geometric Distribution with Time-Varying Covariates}
\description{
Implements the Grassia(II)-Geometric (G2G) distribution model with time-varying covariates for discrete-time survival analysis. Provides maximum likelihood estimation following Fader and Hardie's methodology for incorporating covariates in discrete-time duration models. Supports both static and time-varying covariates using person-period formatted data.
}
\author{
\strong{Maintainer}: Aaron Ramsey \email{aarondaelramsey@gmail.com}

Authors:
\itemize{
  \item Ka Lok Lee \email{kaloklee@gmail.com} [copyright holder]
  \item Noah Park \email{noahpark@wharton.upenn.edu} [contributor]
}

}
\keyword{internal}
